Ttl info move to the base class Inode
