+++ core/server/src/main/java/alluxio/master/file/meta/Inode.java
@@ -15,6 +15,7 @@ import alluxio.Constants;
 import alluxio.master.journal.JournalEntryRepresentable;
 import alluxio.security.authorization.Permission;
 import alluxio.wire.FileInfo;
+import alluxio.wire.TtlAction;
 
 import com.google.common.base.Objects;
 
@@ -34,6 +35,8 @@ public abstract class Inode<T> implements JournalEntryRepresentable {
   private boolean mDeleted;
   protected final boolean mDirectory;
   protected final long mId;
+  protected long mTtl;
+  protected TtlAction mTtlAction;
   private long mLastModificationTimeMs;
   private String mName;
   private long mParentId;
@@ -52,6 +55,8 @@ public abstract class Inode<T> implements JournalEntryRepresentable {
     mDirectory = isDirectory;
     mGroup = "";
     mId = id;
+    mTtl = Constants.NO_TTL;
+    mTtlAction = TtlAction.DELETE;
     mLastModificationTimeMs = mCreationTimeMs;
     mName = null;
     mParentId = InodeTree.NO_PARENT;
@@ -81,6 +86,20 @@ public abstract class Inode<T> implements JournalEntryRepresentable {
    */
   public long getId() {
     return mId;
+  }
+
+  /**
+   * @return the ttl of the file
+   */
+  public long getTtl() {
+    return mTtl;
+  }
+
+  /**
+   * @return the {@link TtlAction}
+   */
+  public TtlAction getTtlAction() {
+    return mTtlAction;
   }
 
   /**
@@ -215,6 +234,24 @@ public abstract class Inode<T> implements JournalEntryRepresentable {
   }
 
   /**
+   * @param ttl the TTL to use, in milliseconds
+   * @return the updated object
+   */
+  public T setTtl(long ttl) {
+    mTtl = ttl;
+    return getThis();
+  }
+
+  /**
+   * @param ttlAction the {@link TtlAction} to use
+   * @return the updated options object
+   */
+  public T setTtlAction(TtlAction ttlAction) {
+    mTtlAction = ttlAction;
+    return getThis();
+  }
+
+  /**
    * @param persistenceState the {@link PersistenceState} to use
    * @return the updated object
    */
@@ -338,6 +375,7 @@ public abstract class Inode<T> implements JournalEntryRepresentable {
   protected Objects.ToStringHelper toStringHelper() {
     return Objects.toStringHelper(this).add("id", mId).add("name", mName).add("parentId", mParentId)
         .add("creationTimeMs", mCreationTimeMs).add("pinned", mPinned).add("deleted", mDeleted)
+        .add("ttl", mTtl).add("mTtlAction", mTtlAction)
         .add("directory", mDirectory).add("persistenceState", mPersistenceState)
         .add("lastModificationTimeMs", mLastModificationTimeMs).add("owner", mOwner)
         .add("group", mGroup).add("permission", mMode);
